pose-angle-api-master/requirements.txt
opencv-python
mediapipe
streamlit
google-generativeai